First test

+++ deploy.sh
@@ -15,11 +15,15 @@ git config --global push.default simple
 git config user.name "Travis CI"
 git config user.email "deploy@travis-ci.org"
 
-echo '[deploy_gh-pages]: remove all old files'
-rm -rf *
-
 echo '[deploy_gh-pages]: copy new files'
-cp ../../output/* .
+if [ "$TRAVIS_BRANCH" == "master" ] then
+    rm -f *
+    cp ../../output/* .
+else
+    rm -rf "$TRAVIS_BRANCH"
+    mkdir -p "$TRAVIS_BRANCH"
+    cp ../../output/* ./"$TRAVIS_BRANCH"/
+fi
 
 echo '[deploy_gh-pages]: git add & commit'
 git add --all
@@ -32,29 +36,32 @@ echo '[deploy_gh-pages]: remove temp dir'
 cd ../..
 rm -rf deployGH-PAGES
 
-#deploy to commplete repo
-echo '[deploy-complete]: Get data'
-commitHash="$1"
-commitMessage=$(git --no-pager log -1 --pretty="%B")
-commitAuthorEMail=$(git --no-pager log -1 --pretty="%cE")
-commitAuthorName=$(git --no-pager log -1 --pretty="%aN")
-echo "[deploy-complete]: Last commit: $commitHash by $commitAuthorName with email $commitAuthorEMail"
 
-# Remove all output, we don't want to commit binaries
-echo '[deploy-complete]: Clone'
-rm -rf output/
-git clone "https://${GH_REPO_TOKEN}@github.com/SoPra-Team-10/Complete.git"
-cd Complete/
-git config --global push.default simple
-git config user.name "$commitAuthorName"
-git config user.email "$commitAuthorEMail"
-echo '[deploy-complete]: Copy'
-rm -rf Uebungsblaetter
-mkdir -p Uebungsblaetter
-cp ../README.md Uebungsblaetter/
-cp -r ../Meilenstein* Uebungsblaetter/
-cp -r ../.gitignore Uebungsblaetter/
-echo '[deploy-complete]: Commit'
-git add -A
-git commit -m "$commitMessage"
-git push -f  "https://${GH_REPO_TOKEN}@github.com/SoPra-Team-10/Complete.git" master
+#deploy to commplete repo, only on master
+if [ "$TRAVIS_BRANCH" == "master" ] then
+    echo '[deploy-complete]: Get data'
+    commitHash="$1"
+    commitMessage=$(git --no-pager log -1 --pretty="%B")
+    commitAuthorEMail=$(git --no-pager log -1 --pretty="%cE")
+    commitAuthorName=$(git --no-pager log -1 --pretty="%aN")
+    echo "[deploy-complete]: Last commit: $commitHash by $commitAuthorName with email $commitAuthorEMail"
+
+    # Remove all output, we don't want to commit binaries
+    echo '[deploy-complete]: Clone'
+    rm -rf output/
+    git clone "https://${GH_REPO_TOKEN}@github.com/SoPra-Team-10/Complete.git"
+    cd Complete/
+    git config --global push.default simple
+    git config user.name "$commitAuthorName"
+    git config user.email "$commitAuthorEMail"
+    echo '[deploy-complete]: Copy'
+    rm -rf Uebungsblaetter
+    mkdir -p Uebungsblaetter
+    cp ../README.md Uebungsblaetter/
+    cp -r ../Meilenstein* Uebungsblaetter/
+    cp -r ../.gitignore Uebungsblaetter/
+    echo '[deploy-complete]: Commit'
+    git add -A
+    git commit -m "$commitMessage"
+    git push -f  "https://${GH_REPO_TOKEN}@github.com/SoPra-Team-10/Complete.git" master
+fi
